% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default}
\alias{default}
\title{Federal Default}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.}
\source{
\url{https://www.cms.gov/CCIIO/Programs-and-Initiatives/Health-Insurance-Market-Reforms/state-rating.html}
}
\usage{
default
}
\description{
Health Insurance Market Rating Reform: Federal default parameters.
}
\keyword{datasets}

