% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgm}
\alias{sgm}
\title{Small Group Market: State Specific Rating Variations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 9 columns.}
\source{
\url{https://www.cms.gov/CCIIO/Programs-and-Initiatives/Health-Insurance-Market-Reforms/state-rating.html}
}
\usage{
sgm
}
\description{
A data frame contains state specific rating variations for small group health insurance market, specified by Health Insurance Market Reform.
}
\keyword{datasets}

